report/requirements.txt
requests
aws-lambda-powertools[all]
atlassian-python-api
Jinja2
pydantic